Added module stormpy.exceptions
