+++ lib/stormpy/exceptions/__init__.py
@@ -1,11 +1,1 @@
-class StormError(Exception):
-    """
-    Base class for exceptions in Storm.
-    """
-
-    def __init__(self, message):
-        """
-        Constructor.
-        :param message: Error message.
-        """
-        self.message = "Storm: " + message
+from .storm_error import StormError
